[language server] refactor class structure
