+++ packages/argdown-language-server/package.json
@@ -44,5 +44,9 @@
     "vscode-languageserver": "^9.0.1",
     "vscode-languageserver-textdocument": "1.0.12"
   },
-  "packageManager": "yarn@4.9.4"
+  "packageManager": "yarn@4.9.4",
+  "dependencies": {
+    "fast-glob": "^3.3.3",
+    "vscode-uri": "^3.1.0"
+  }
 }
